ai-service/requirements.txt
fastapi
uvicorn
neo4j
chromadb
supabase
pydantic
pydantic-settings
google-generativeai
python-dotenv
asyncio
httpx
datasets
pandas
tqdm
